Change score calculation to double, and removal of FileScores since it was implemented wrong

+++ plugin/mutate/source/dextool/plugin/mutate/backend/report/html/package.d
@@ -566,7 +566,7 @@ void toIndex(FileIndex[] files, Element root, string htmlFileDir, FileScore[] sc
             }();
             r.addChild("td", format!"%.3s"(score)).style = style;
 
-            float scoreChange;
+            double scoreChange;
             if(f.display in scoreDifference){
               scoreChange = scoreDifference[f.display];
             }else{
